add package URP and add pixelate test scene
will add :
- URP Package
- Test folder with test scene and assets

Will show an exemple of pixelization on running time.

to debug:
make background of pixelated object texture transparent.  currently, we pixelize background with the object but is is awful(ly close to what we want but not what we want)
This is managed in PixelateRenderTexture file wich is updated by pixel cam every frame